requirements.txt
numpy
pydicom
Pillow